refector: response.ts 에서 관리되던 모든 타입들을 도메인별로 분리
