+++ src/types/api/channel.ts
@@ -1,0 +1,7 @@
+import type DefaultResponse from './default';
+
+export interface ChannelResponse extends DefaultResponse {
+  posts: string[];
+  name: string;
+  description: string;
+}
