+++ src/types/api/default/index.ts
@@ -1,0 +1,5 @@
+export default interface DefaultResponse {
+  _id: string;
+  createdAt: string;
+  updatedAt?: string;
+}
