+++ src/types/api/like.ts
@@ -1,0 +1,6 @@
+import type ApiDefaultResponse from './default';
+
+export interface LikeResponse extends ApiDefaultResponse {
+  user: string; // 사용자 id
+  post: string; // 포스트 id
+}
