+++ src/types/recoil/user.ts
@@ -1,0 +1,13 @@
+import type { UserResponse } from '../api/user';
+
+export type RecoilUser = Pick<
+  UserResponse,
+  | '_id'
+  | 'image'
+  | 'fullName'
+  | 'likes'
+  | 'comments'
+  | 'email'
+  | 'following'
+  | 'notifications'
+>;
